Plugin Directory: API: Cast the `rating` field to an int just incase `\WPORG_Ratings::get_avg_rating()` were to ever return a value that doesn't map to a round percentage.
While `::get_avg_rating()` has been fixed, this caused some API clients such as the mobile apps to hit unexpected values.

See https://github.com/wordpress-mobile/WordPress-iOS/issues/16056


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@10816 74240141-8908-4e6f-9713-ba540dce6ec7

+++ wordpress.org/public_html/wp-content/plugins/plugin-directory/api/routes/class-plugin.php
@@ -124,7 +124,7 @@ class Plugin extends Base {
 			$result['ratings'] = get_post_meta( $post->ID, 'ratings', true ) ?: array();
 		}
 
-		$result['rating']  = $result['rating'] * 20; // Stored as 0.0 ~ 5.0, API outputs as 0..100
+		$result['rating']  = (int) ( $result['rating'] * 20 ); // Stored as 0.0 ~ 5.0, API outputs as 0..100
 		$result['ratings'] = array_map( 'intval', $result['ratings'] );
 		krsort( $result['ratings'] );
